requirements.txt
crispy-bootstrap5==0.7
Django==4.1.7
django-rq==2.7.0
djangorestframework==3.14.0
psycopg2-binary==2.9.5

Pillow==9.4.0

pytest>=7.2.2
pytest-django>=4.5.2
scrapy==2.8.0

redis==4.5.4



factory-boy>=3.2.1
pytest-factoryboy>=2.5.0
faker>=13.15.0

gunicorn==20.1.0


flake8>=6.0.0
isort>=5.12.0
black>=23.3.0
